refactor(layout): split table row heuristics
Table detection had grown into one file covering candidate trimming, continuation rows, leading headers, and column recurrence. Split those responsibilities into focused modules so future table fixes can land without expanding the central detector.
